requirements.txt
chainlit
openai
python-dotenv
langchain
langchain-openai
chromadb
firebase-admin
gunicorn